Refactors Intersection namespaces. Adds placeholder folders.
