+++ src/GShark.Test.XUnit/Operation/IntersectionTests.cs
@@ -1,9 +1,9 @@
 ﻿using FluentAssertions;
 using GShark.Core;
-using GShark.Core.IntersectionResults;
 using GShark.Geometry;
 using GShark.Operation;
 using System.Collections.Generic;
+using GShark.Intersection;
 using Xunit;
 using Xunit.Abstractions;
