+++ src/GShark/Geometry/Circle.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Linq;
 using GShark.Interfaces;
+using GShark.Intersection;
 using GShark.Operation;
 
 namespace GShark.Geometry
